Use OSFTSAVE for both saves and restores
Originally OSFTSAVE was assumed, but then if neither "restore" nor
"game" was found in the prompt, it was switched to OSFTUNK, meaning when
saving a file, Glk would be told the file usage is data, whereas for
saving, it was told the file usage is save. That resulted in save
dialogs filtering on different extension from restore dialogs.

Now OSFTSAVE is used when "save" OR "restore" is present along with
"game".